% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hreg_bayes.R
\name{amsprod}
\alias{amsprod}
\title{Array matrix slice product}
\usage{
amsprod(X, B)
}
\arguments{
\item{X}{An \code{r} by \code{p} by \code{n} array.}

\item{B}{An \code{r} by \code{p} matrix.}
}
\value{
An \code{r} by \code{n} matrix, obtained by
matrix multiplying each row of \code{B} by the
matrix obtained from the \code{r}th first-mode
slice of \code{X}.
}
\description{
Multiply matrices and vectors that
correspond to the same slices as
an array and matrix, respectively.
}
\author{
Peter Hoff
}

