% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ranks.R
\name{kendalltau}
\alias{kendalltau}
\title{Kendall's tau measure of association}
\usage{
kendalltau(x, y, nmc = 1e+05)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}

\item{nmc}{an integer number of Monte Carlo simulations}
}
\value{
An array of dimension \code{dim}.
}
\description{
This function provides a Monte Carlo approximation to
Kendall's tau measure of association.
}
\examples{
rsan(c(5,4,3))
}
\author{
Peter Hoff
}

