% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hreg_bayes.R
\name{rb0_fc}
\alias{rb0_fc}
\title{Bayes Hierarchical Linear Regression Posterior Simulation}
\usage{
rb0_fc(B, V0, iV0 = solve(V0), eb0 = rep(0, ncol(B)), iVb0 = diag(nrow =
  ncol(B))/nrow(B))
}
\arguments{
\item{B}{An \code{r} by \code{p} matrix}

\item{V0}{An \code{mp} by \code{mp} matrix.}

\item{iV0}{The inverse of \code{V0}.}
}
\value{
A \code{p} vector simulated
from the posterior distribution.
}
\description{
This functions returns a simulation from the posterior distribution
of the across-group mean of the regression vectors.
}
\author{
Peter Hoff
}

