% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hreg_bayes.R
\name{rB_fc}
\alias{rB_fc}
\title{Bayes Hierarchical Linear Regression Posterior Simulation}
\usage{
rB_fc(Y, X, b0, V0, ve = 1)
}
\arguments{
\item{Y}{An \code{r} by \code{n} matrix}

\item{X}{An \code{r} by \code{p} by \code{n} array.}

\item{b0}{A \code{p} vector.}

\item{V0}{An \code{p} by \code{p} matrix.}

\item{ve}{A scalar.}
}
\value{
An \code{r} by \code{p} matrix, simulated
from the posterior distribution.
}
\description{
This functions returns a simulation from the posterior distribution
of the matrix of linear regression coefficients.
}
\author{
Peter Hoff
}

