% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rsan.R
\name{rsan}
\alias{rsan}
\title{Random Standard Normal Array}
\usage{
rsan(dim)
}
\arguments{
\item{dim}{A vector of positive integers.}
}
\value{
An array of dimension \code{dim}.
}
\description{
This functions generates an array of dimension \code{dim}
filled with iid standard normal random variables.
}
\examples{
rsan(c(5,4,3))
}
\author{
Peter Hoff
}

