% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hreg_bayes.R
\name{rV0_fc}
\alias{rV0_fc}
\title{Bayes Hierarchical Linear Regression Posterior Simulation}
\usage{
rV0_fc(B, b0, S0 = diag(nrow = length(b0)), nu0 = length(b0) + 1)
}
\arguments{
\item{B}{An \code{r} by \code{p} array.}

\item{b0}{A \code{p} vector.}
}
\value{
An \code{p} by \code{p} matrix simulated
from the posterior distribution.
}
\description{
This functions returns a simulation from the posterior distribution
of the across-group covariance of the regression vectors.
}
\author{
Peter Hoff
}

