% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rwish.R
\name{rwish}
\alias{rwish}
\title{Wishart simulation}
\usage{
rwish(S0, nu = dim(as.matrix(S0))[1] + 1)
}
\arguments{
\item{S0}{A positive definite matrix.}

\item{nu}{A positive scalar}
}
\value{
A square matrix.
}
\description{
Simulate a Wishart-distributed random matrix
using Bartletts decomposition, as described in Everson and
Morris (2000).
}
\examples{
# simulate several matrices and compute the mean
SS<-matrix(0,5,5)
for(s in 1:1000) { SS<-SS+rwish(diag(5),3) }
SS/s
}
\author{
Peter Hoff
}

