% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ranks.R
\name{rz_fc}
\alias{rz_fc}
\title{Gibbs sampling for the Gaussian rank likelihood}
\usage{
rz_fc(y, z, ez, sz = 1)
}
\arguments{
\item{y}{A vector, matrix or array.}

\item{z}{A vector, matrix or array of the same dimension as \code{y}.}

\item{ez}{A vector, matrix or array of the same dimension as \code{y}.}

\item{sz}{A scalar, the common standard deviation of \code{z}.}
}
\value{
A vector, matrix or array of the same dimension as \code{y}.
}
\description{
This function simulates a latent Gaussian object
\code{z} conditional on a partial ordering induced by
an observed data object \code{y}.
}
\author{
Peter Hoff
}

